refactor: separate components, views, and styles
