+++ src/components/impact-card/index.jsx
@@ -1,6 +1,6 @@
 import Image from "next/image"
 import Link from "next/link"
-import cardStyles from "../styles/impactCard.module.css"
+import cardStyles from "./index.module.css"
 import { Icon } from "@iconify/react"
 
 export default function ImpactCard({ title, imageUrl, url }) {
